requirements.txt
dnspython==2.7.0
greenlet==3.2.3
numpy==2.3.1
pandas==2.3.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sqlalchemy==2.0.41
typing-extensions==4.14.1
tzdata==2025.2
cffi==1.17.1
cryptography==45.0.5
dnspython==2.7.0
faker==37.4.0
greenlet==3.2.3
loguru==0.7.3
numpy==2.3.1
pandas==2.3.1
pycparser==2.22
pymongo==4.13.2
pymysql==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sqlalchemy==2.0.41
tqdm==4.67.1
typing-extensions==4.14.1
tzdata==2025.2
